pyproject.toml
[project]
name = "proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "torch>=2.9.1",
    "scikit-learn>=1.5.0",
    "matplotlib>=3.8.0",
    "torchvision>=0.16.1",
    "dvc>=3.65.0",
    "dvc-gdrive>=3.0.1",
    "pathspec<0.12.0",
    "hydra-core>=1.3.2",
]

[dependency-groups]
dev = [
    "coverage>=7.13.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.ruff]
lint.select = ["I"]
